+++ docs/conf.py
@@ -25,7 +25,7 @@ copyright = '2018, Laurent LAPORTE'
 author = 'Laurent LAPORTE'
 
 # The full version, including alpha/beta/rc tags
-release = '0.4.0'
+release = '0.4.1'
 # The short X.Y version
 version = '.'.join(release.split('.')[:2])