+++ CHANGELOG.rst
@@ -8,6 +8,12 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
 
+v0.4.1 (unreleased)
+===================
+
+Bug fix release
+
+
 v0.4.0 (2019-04-23)
 ===================